Split legacy endpoints to a different module; partial pylint
