+++ src/bayestourney/app.py
@@ -6,31 +6,22 @@ Created on Jun 4, 2013
 @author: welling
 '''
 
-import io
-import time
 import json
 import logging
-import functools
-from math import ceil
+from pathlib import Path
+from pprint import pprint
 
 from flask import (
-    Blueprint, flash, g, render_template, request, url_for,
-    send_from_directory, send_file, session, redirect, current_app
+    Blueprint, flash, render_template, request, url_for,
+    send_file, session, redirect, current_app
 )
 from flask_login import login_required, current_user
 from werkzeug.utils import secure_filename
 from werkzeug.routing import BuildError
-from werkzeug.exceptions import abort
-from sqlalchemy.sql import text as sql_text
 from sqlalchemy.exc import NoResultFound
-import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
-from matplotlib.backends.backend_svg import FigureCanvasSVG as FigureCanvas
-from matplotlib.figure import Figure
-from pathlib import Path
-from pprint import pprint
 
+from .wrappers import debug_wrapper, debug_page_wrapper
 from .database import get_db
 from .models import (Tourney, LogitPlayer, Bout, User,
                      Group, TourneyPlayerPair, DBException)
@@ -43,6 +34,7 @@ from .permissions import (
     current_user_can_read, current_user_can_write,
     )
 from . import stat_utils
+from .legacy_endpoints import bp as legacy_bp, _get_bearpit_dataframe
 
 
 UPLOAD_ALLOWED_EXTENSIONS = ['tsv', 'csv']
@@ -52,7 +44,9 @@ logging.basicConfig()
 LOGGER = logging.getLogger(__name__)
 LOGGER.setLevel(logging.DEBUG)
 
+
 bp = Blueprint('', __name__)
+bp.register_blueprint(legacy_bp)
 
 
 @bp.app_errorhandler(PermissionException)
@@ -67,45 +61,13 @@ def handle_db_exception(e):
     return f"{e}", 403
 
 
-def logMessage(txt):
-    LOGGER.info(txt)
-
-
-def debug_wrapper(view):
-    @functools.wraps(view)
-    def wrapped_view(**kwargs):
-        LOGGER.debug(f'{request.method} Request for {request.endpoint}'
-                     f' kwargs {kwargs}'
-                     f' params {[(k,request.values[k]) for k in request.values]}')
-        rslt = view(**kwargs)
-        LOGGER.debug(f'Returning {rslt}')
-        return rslt
-    return wrapped_view
-
-    
-def debug_page_wrapper(view):
-    @functools.wraps(view)
-    def wrapped_view(**kwargs):
-        LOGGER.debug(f'{request.method} Request for {request.endpoint}'
-                     f' kwargs {kwargs}'
-                     f' params {[(k,request.values[k]) for k in request.values]}'
-        )
-        #print('session state before view follows:')
-        #pprint(session)
-        rslt = view(**kwargs)
-        #print('session follows:')
-        #pprint(session)
-        return rslt
-    return wrapped_view
-
-    
 def allowed_upload_file(filename):
     return ('.' in filename and
             Path(filename).suffix[1:] in UPLOAD_ALLOWED_EXTENSIONS)
 
 
 def insert_entrants_from_df(df):
-    col_set = set([str(col) for col in df.columns])
+    col_set = set(str(col) for col in df.columns)
     db = get_db()
     if set(["name", "note"]).issubset(col_set):
         df["id"] = df.apply(player_name_to_id,
@@ -182,12 +144,12 @@ def upload_entrants_file():
 def player_name_to_id(row, key, db):
     try:
         return db.query(LogitPlayer).filter_by(name=row[key]).one().id
-    except:
+    except NoResultFound:
         return -1  # we need to use an int as the signal value for Pandas' sake
 
 
 def insert_bouts_from_df(df, tourney):
-    col_set = set([str(col) for col in df.columns])
+    col_set = set(str(col) for col in df.columns)
     db = get_db()
     if set(["leftWins", "leftPlayerName", "rightPlayerName",
             "rightWins", "draws"]).issubset(col_set):
@@ -202,7 +164,7 @@ def insert_bouts_from_df(df, tourney):
                                        axis=1)
         bad_names = df[df['leftPlayerId'] == -1]['leftPlayerName']
         bad_names = bad_names.append(df[df['rightPlayerId'] == -1]['rightPlayerName'])
-        if bad_name_l := [nm for nm in bad_names.unique()]:
+        if bad_name_l := list(bad_names.unique()):
             bad_name_str = ', '.join([f'"{nm}"' for nm in bad_name_l])
             raise DBException(f'Unknown or multiply defined player names: {bad_name_str}')
         bad_names = []
@@ -232,7 +194,7 @@ def insert_bouts_from_df(df, tourney):
                             note = note)
             db.add(new_bout)
         db.commit()
-                         
+
     else:
         raise DBException('Unknown column pattern for bouts')
 
@@ -281,8 +243,8 @@ def upload_bouts_file():
     LOGGER.info('Parse complete')
     try:
         tourney = get_db().query(Tourney).filter_by(tourneyId=tourney_id).one()
-    except NoResultFound:
-        raise RuntimeError("No such tourney")
+    except NoResultFound as exc:
+        raise RuntimeError("No such tourney") from exc
     LOGGER.info(f'Tourney is {tourney}')
     check_can_write(tourney)
     try:
@@ -297,12 +259,12 @@ def upload_bouts_file():
 
 @bp.route("/forms/entrants/create")
 def forms_entrants_create():
-    return render_template("forms_entrants_create.html");
+    return render_template("forms_entrants_create.html")
 
 
 @bp.route("/forms/tourneys/create")
 def forms_tourneys_create():
-    return render_template("forms_tourneys_create.html");
+    return render_template("forms_tourneys_create.html")
 
 
 @bp.route("/site-map")
@@ -318,9 +280,7 @@ def site_map():
                 url = f'BuildError {e}'
             links.append((url, rule.endpoint))
     # links is now a list of url, endpoint tuples
-    return render_template("site_map.html",
-                           link_dict={k:v for k, v in links})
-    print(links)
+    return render_template("site_map.html", link_dict=dict(links))
 
 
 @bp.route("/settings")
@@ -336,20 +296,20 @@ def admin_page():
 @bp.route('/')
 @login_required
 def index():
-    return redirect(url_for('tourneys'))
+    return redirect(url_for('tourneys_fun'))
 
 
 @bp.route('/tourneys')
 @login_required
 @debug_page_wrapper
-def tourneys():
+def tourneys_fun():
     return render_template("tourneys.html")
 
 
 @bp.route('/entrants')
 @login_required
 @debug_page_wrapper
-def entrants():
+def entrants_fun():
     if 'tourney_id' in request.values:
         tourney_id = int(request.values['tourney_id'])
         session['sel_tourney_id'] = tourney_id
@@ -362,7 +322,7 @@ def entrants():
 @bp.route('/bouts')
 @login_required
 @debug_page_wrapper
-def bouts():
+def bouts_fun():
     db = get_db()
     if 'tourney_id' in request.values:
         tourney_id = int(request.values['tourney_id'])
@@ -382,248 +342,26 @@ def bouts():
                            playerDict=playerDict)
 
 
-@bp.route('/experiment')
-#@login_required
-@debug_page_wrapper
-def experiment():
-    session.pop('horserace_includes')
-    get_db().commit()
-    return 'the experiment happened'
-
-
 @bp.route('/horserace')
 @login_required
 @debug_page_wrapper
-def horserace():
+def horserace_fun():
     tourneyDict = {t.tourneyId: t.name for t in get_readable_tourneys(get_db())}
     return render_template("horserace.html",
                            sel_tourney_id=session.get('sel_tourney_id', -1),
                            tourneyDict=tourneyDict)
 
 
-@bp.route('/hello')
-def hello():
-    return "Hello World"
-
-
 @bp.route('/help')
 @debug_page_wrapper
-def help():
-    return render_template("info.html")    
+def help_fun():
+    return render_template("info.html")
 
 
 @bp.route('/notimpl')
 @debug_page_wrapper
 def notimplPage():
-    return flask.static_file('notimpl.html', root='../../www/static/')
-
-
-def _orderAndChopPage(pList,fieldMap):
-    """
-    This was very useful for the old version of jqGrid, before 'loadonce: true'
-    """
-    sortIndex = request.values['sidx']
-    sortOrder = request.values['sord']
-    thisPageNum = int(request.values['page'])
-    rowsPerPage = int(request.values['rows'])
-    if sortIndex in fieldMap:
-        field = fieldMap[sortIndex]
-        pDict = {(getattr(p, field), idx) : p for idx, p in enumerate(pList)}
-        sortMe = [tpl for tpl in pDict]
-        if sortOrder == 'asc':
-            sortMe.sort()
-        else:
-            sortMe.sort(reverse=True)
-        pList = [pDict[tpl] for tpl in sortMe]
-        nPages = int(ceil(float(len(pList))/(rowsPerPage-1)))
-        totRecs = len(pList)
-        if thisPageNum == nPages:
-            eR = totRecs
-            sR = max(eR - rowsPerPage, 0)
-        else:
-            sR = (thisPageNum-1)*(rowsPerPage-1)
-            eR = sR + rowsPerPage
-        pList = pList[sR:eR]
-        return (nPages,thisPageNum,totRecs,pList)
-    else:
-        raise RuntimeError("Sort index %s not in field map"%sortIndex)
-
-@bp.route('/list/<path>')
-@debug_page_wrapper
-def handleList(path):
-    """
-    Used for populating select elements
-    returns html
-    """
-    db = get_db()
-    if path=='select_entrant':
-        playerList = db.query(LogitPlayer)
-        pairs = [(p.id,p.name) for p in playerList]
-        pairs.sort()
-        s = "<select>\n"
-        for thisId,name in pairs:
-            s += "<option value=%d>%s<option>\n"%(thisId,name)
-        s += "</select>\n"
-        return s
-    elif path=='select_tourney':
-        tourneyList = get_readable_tourneys(db)
-        pairs = [((t.tourneyId,t.name)) for t in tourneyList]
-        pairs.sort()
-        s = "<select>\n"
-        for thisId,name in pairs: s += "<option value=%d>%s<option>\n"%(thisId,name)
-        s += "</select>\n"
-        return s
-    else:
-        raise RuntimeError("Bad path /list/%s"%path)
-    
-
-@bp.route('/edit/<path>', methods=['POST'])
-@debug_wrapper
-def handleEdit(path):
-    """
-    Specialized endpoint that understands the requests jqgrid uses to perform
-    edits, adds, and deletes.
-    """
-    db = get_db()
-    engine = db.get_bind()
-    if path=='tourneys':
-        if request.values['oper']=='edit':
-            try:
-                t = db.query(Tourney).filter_by(tourneyId=int(request.values['id'])).one()
-            except NoResultFound:
-                raise RuntimeError("No such tourney")
-            check_can_write(t)
-            if 'name' in request.values:
-                t.name = request.values['name']
-            if 'notes' in request.values:
-                t.note = request.values['notes']
-            db.commit()
-            return {}
-        elif request.values['oper']=='add':
-            name = request.values['name']
-            notes = request.values['notes']
-            if db.query(Tourney).filter_by(name=name).count() != 0:
-                raise RuntimeError('There is already a tourney named %s'%name)
-            current_user_group = (db.query(Group)
-                                  .filter_by(name=current_user.username)
-                                  .one())
-            t = Tourney(name, current_user.id, current_user_group.id, notes)
-            check_can_write(t)
-            db.add(t)
-            db.commit()
-            return {}
-        elif request.values['oper']=='del':
-            tourneyId = int(request.values['id'])
-            try:
-                tourney = db.query(Tourney).filter_by(tourneyId=tourneyId).one()
-            except NoResultFound:
-                raise RuntimeError("no such tourney")
-            check_can_delete(tourney)
-            bouts = db.query(Bout).filter_by(tourneyId=tourneyId)
-            for bout in bouts:
-                db.delete(bout)
-            db.delete(tourney)
-            db.commit()
-            return {}
-        else:
-            raise RuntimeError(f"Bad edit operation {request.values['oper']}")
-    elif path=='bouts':
-        if request.values['oper']=='edit':
-            try:
-                b = db.query(Bout).filter_by(boutId=int(request.values['id'])).one()
-            except NoResultFound:
-                raise RuntimeError("no such bout")
-            tourney = db.query(Tourney).filter_by(tourneyId=b.tourneyId).one()
-            check_can_write(tourney)
-            if 'tourney' in request.values:
-                b.tourneyId = int(request.values['tourney'])
-            if 'rightplayer' in request.values:
-                b.rightPlayerId = int(request.values['rightplayer'])
-            if 'leftplayer' in request.values:
-                b.leftPlayerId = int(request.values['leftplayer'])
-            if 'rwins' in request.values:
-                b.rightWins = int(request.values['rwins'])
-            if 'lwins' in request.values:
-                b.leftWins = int(request.values['lwins'])
-            if 'draws' in request.values:
-                b.draws = int(request.values['draws'])
-            if 'notes' in request.values:
-                b.note = request.values['notes']
-            db.commit()
-            return {}
-        elif request.values['oper']=='add':
-            tourneyId = int(request.values['tourney'])
-            tourney = db.query(Tourney).filter_by(tourneyId=tourneyId).one()
-            check_can_write(tourney)
-            lPlayerId = int(request.values['leftplayer'])
-            rPlayerId = int(request.values['rightplayer'])
-            lWins = int(request.values.get('lwins', '0'))
-            rWins = int(request.values.get('rwins', '0'))
-            draws = int(request.values.get('draws', '0'))
-            note = request.values.get('notes', '')
-            b = Bout(tourneyId, lWins, lPlayerId, draws,
-                     rPlayerId, rWins, note)
-            db.add(b)
-            db.commit()
-            return {}
-        elif request.values['oper']=='del':
-            try:
-                b = db.query(Bout).filter_by(boutId=int(request.values['id'])).one()
-            except NoResultFound:
-                raise RuntimeError(f"No such bout")
-            tourney = db.query(Tourney).filter_by(tourneyId=b.tourneyId).one()
-            check_can_write(tourney)
-            db.delete(b)
-            db.commit()
-            return {}
-        else:
-            raise RuntimeError(f"Bad edit operation {request.values['oper']}")
-    elif path=='entrants':
-        if request.values['oper']=='edit':
-            try:
-                p = db.query(LogitPlayer).filter_by(id=int(request.values['id'])).one()
-            except NoResultFound:
-                raise RuntimeError("No such entrant")
-            if 'name' in request.values:
-                p.name = request.values['name']
-            if 'notes' in request.values:
-                p.note = request.values['notes']
-            db.commit()
-            return {}
-        elif request.values['oper']=='add':
-            name = request.values['name']
-            notes = request.values['notes']
-            LogitPlayer.create_unique(db, name, notes)
-            db.commit()
-            return {}
-        elif request.values['oper'] == 'del':
-            player_id = int(request.values['id'])
-            try:
-                player = db.query(LogitPlayer).filter_by(id=player_id).one()
-            except NoResultFound:
-                raise RuntimeError("No such entrant")
-            with engine.connect() as conn:
-                stxt = sql_text(
-                    """
-                    select * from bouts
-                    where bouts.leftPlayerId = :p_id or bouts.rightPlayerId = :p_id
-                    """
-                )
-                rs = conn.execute(stxt, p_id=player_id)
-                num_bouts = len([rec for rec in rs])
-            if num_bouts == 0:
-                db.delete(player)
-                db.commit()
-                return {"status":"success"}
-            else:
-                return {"status":"failure",
-                        "msg":("The player could not be deleted because there"
-                               " are bouts which include them")
-                        }
-        else:
-            raise RuntimeError(f"Bad edit operation {request.values['oper']}")
-    else:
-        raise RuntimeError("Bad path /edit/%s"%path)
+    return render_template('notimpl.html')
 
 
 def _get_bouts_dataframe(tourneyId: int, include_ids: bool = False) -> pd.DataFrame:
@@ -638,7 +376,7 @@ def _get_bouts_dataframe(tourneyId: int, include_ids: bool = False) -> pd.DataFr
         bouts = db.query(Bout).filter(Bout.tourneyId.in_(tourney_id_list)).all()
     dict_list = []
     for bout in bouts:
-        bout_tourney = db.query(Tourney).filter_by(tourneyId=bout.tourneyId).one();
+        bout_tourney = db.query(Tourney).filter_by(tourneyId=bout.tourneyId).one()
         dct = {'tourneyName': bout_tourney.name,
                'leftWins': bout.leftWins,
                'leftPlayerName': bout.lName,
@@ -648,72 +386,12 @@ def _get_bouts_dataframe(tourneyId: int, include_ids: bool = False) -> pd.DataFr
                'note': bout.note}
         if include_ids:
             dct['bout_id'] = bout.boutId
-            dct['tourney_id'] = bout.tourneyId,
+            dct['tourney_id'] = bout.tourneyId
             dct['leftPlayerId'] = bout.leftPlayerId
             dct['rightPlayerId'] = bout.rightPlayerId
         dict_list.append(dct)
     boutDF = pd.DataFrame(dict_list)
     return boutDF
-
-
-def _get_bearpit_dataframe(db, bouts):
-    dict_list = []
-    for bout in bouts:
-        bout_tourney = db.query(Tourney).filter_by(tourneyId=bout.tourneyId).one();
-        dct = {'tourneyName': bout_tourney.name,
-               'leftWins': bout.leftWins,
-               'leftPlayerName': bout.lName,
-               'draws': bout.draws,
-               'rightPlayerName': bout.rName,
-               'rightWins': bout.rightWins,
-               'note': bout.note,
-               'bout_id': bout.boutId,
-               'tourney_id': bout.tourneyId,
-               'leftPlayerId': bout.leftPlayerId,
-               'rightPlayerId': bout.rightPlayerId
-               }
-        dict_list.append(dct)
-    if (dict_list):
-        boutDF = pd.DataFrame(dict_list)
-        #print(boutDF.head())
-        #print(boutDF.columns)
-
-        leftDF = boutDF[['leftPlayerName', 'leftPlayerId', 'leftWins', 'rightWins',
-                         'draws']].copy()
-        leftDF['bouts'] = leftDF['leftWins'] + leftDF['rightWins'] + leftDF['draws']
-        leftDF = leftDF.rename(columns={'leftWins': 'wins',
-                                        'rightWins': 'losses',
-                                        'leftPlayerName': 'playerName',
-                                        'leftPlayerId': 'id'})
-
-        rightDF = boutDF[['rightPlayerName', 'rightPlayerId', 'rightWins', 'leftWins',
-                          'draws']].copy()
-        rightDF['bouts'] = rightDF['leftWins'] + rightDF['rightWins'] + rightDF['draws']
-        rightDF = rightDF.rename(columns={'leftWins': 'losses',
-                                          'rightWins': 'wins',
-                                          'rightPlayerName': 'playerName',
-                                          'rightPlayerId': 'id'})
-        rightDF = rightDF[leftDF.columns]
-
-        workDF = pd.concat([leftDF, rightDF])
-        workDF = workDF.groupby(['playerName', 'id']).sum().reset_index()
-
-        bearpit_wt_dct = {  # tuples are weights for (wins, losses, draws)
-            'hr_draws_rule_ignore': (2, 1, 0),
-            'hr_draws_rule_win': (2, 1, 2),
-            'hr_draws_rule_loss': (2, 1, 1)
-        }
-        wt_win, wt_loss, wt_draw = bearpit_wt_dct[get_settings()['hr_draws_rule']]
-        workDF['bearpit'] = (wt_win * workDF['wins'] + wt_loss * workDF['losses']
-                             + wt_draw * workDF['draws'])
-        workDF = workDF.sort_values(by='id', axis='rows')
-    else:
-        workDF = pd.DataFrame(columns=['playerName', 'id',
-                                       'wins', 'losses', 'draws',
-                                       'bouts', 'bearpit'])
-    #print(workDF.head())
-    #print(workDF.columns)
-    return workDF
 
 
 @bp.route('/download/bouts')
@@ -728,7 +406,7 @@ def handleBoutsDownloadReq(**kwargs):
     else:
         tourney_name = 'ALL_TOURNEYS'
     boutDF = _get_bouts_dataframe(tourneyId)
-    
+
     session_scratch_dir = current_app.config['SESSION_SCRATCH_DIR']
     full_path =  Path(session_scratch_dir) / f'bouts_{tourney_name}.tsv'
     boutDF.to_csv(full_path, sep='\t', index=False)
@@ -796,37 +474,6 @@ def _horserace_dataframes(db, tourneys, checkboxes):
     return boutDF, playerDF, checkbox_dict
 
 
-def _horserace_fetch_dataframes(data, **kwargs):
-    db = get_db()
-    engine = db.get_bind()
-    tourneyId = int(data['tourney'])
-    if tourneyId >= 0:
-        tourney = db.query(Tourney).filter_by(tourneyId=tourneyId).one()
-        check_can_read(tourney)
-        boutDF = pd.read_sql(f'select * from bouts'
-                             f' where tourneyId={tourneyId}',
-                             engine, coerce_float=True)
-        playerDF = pd.read_sql('select distinct players.*'
-                               ' from players inner join bouts'
-                               ' on ( bouts.leftPlayerId = players.id'
-                               ' or bouts.rightPlayerId = players.id )'
-                               f' where bouts.tourneyId={tourneyId}',
-                               engine, coerce_float=True)
-    else:
-        boutDF = pd.read_sql_table('bouts', engine, coerce_float=True)
-        playerDF = pd.read_sql_table('players', engine, coerce_float=True)
-    checkbox_dict = {int(k): v for k, v in data['checkboxes'].items()}
-    for player_id, flag in checkbox_dict.items():
-        _set_hr_include(tourneyId, player_id, flag)
-    playerDF = playerDF[playerDF.apply(_include_fun, axis=1,
-                                       keycols=['id'], checkbox_dict=checkbox_dict)]
-    boutDF = boutDF[boutDF.apply(_include_fun, axis=1,
-                                 keycols=['leftPlayerId', 'rightPlayerId'],
-                                 checkbox_dict=checkbox_dict)]
-
-    return playerDF, boutDF, checkbox_dict
-
-
 def _get_include_flag(row, tourney_id):
     return _get_hr_include(tourney_id,row['id'])
 
@@ -859,73 +506,6 @@ def handleBearpitDownloadReq(**kwargs):
     bearpitDF.to_csv(full_path, sep='\t', index=False)
     return send_file(full_path, as_attachment=True)
 
-
-@bp.route('/horserace_go', methods=['POST'])
-@debug_page_wrapper
-@login_required
-def horserace_go(**kwargs):
-    data = request.get_json()
-    #logMessage(f"data: {data}")
-    playerDF, boutDF, checkbox_dict = _horserace_fetch_dataframes(data, **kwargs)
-    output = io.StringIO()
-
-    try:
-        fit_info = stat_utils.estimate(
-            playerDF, boutDF,
-            draws_rule=get_settings()['hr_draws_rule']
-        )
-        plt.figure(figsize=[3,3])
-        fig, axes = plt.subplots(ncols=1, nrows=1)
-        graph_yscale_dct = {'hr_graph_yscale_linear': 'linear',
-                            'hr_graph_yscale_log': 'log'}
-        axes.set_yscale(graph_yscale_dct[get_settings()['hr_graph_yscale']])
-        graph_type_dct = {'hr_graph_style_box': 'boxplot',
-                          'hr_graph_style_violin': 'violin'}
-        graph_type = graph_type_dct[get_settings()['hr_graph_style']]
-        fit_info.gen_graph(fig, axes, graph_type)
-        FigureCanvas(fig).print_svg(output)
-        
-    except RuntimeError as e:
-        logMessage('horseRace_go exception: %s' % str(e))
-    result = {'image': output.getvalue(),
-              'announce_html': fit_info.estimate_win_probabilities().as_html()
-              }
-
-    return result
-    
-
-@bp.route('/horserace_get_bouts_graph', methods=['POST'])
-@debug_page_wrapper
-@login_required
-def horserace_get_bouts_graph(**kwargs):
-    data = request.get_json()
-    #logMessage(f"data: {data}")
-    player_df, bouts_df, checkbox_dict = _horserace_fetch_dataframes(
-        data,
-        **kwargs
-    )
-    tourney_id = int(data['tourney'])
-    try:
-        tourney = get_db().query(Tourney).filter_by(tourneyId=tourney_id).one()
-    except NoResultFound:
-        raise RuntimeError("No such tourney")
-    check_can_read(tourney)
-
-    try:
-        fit_info = stat_utils.ModelFit.from_raw_bouts(
-            player_df, bouts_df,
-            draws_rule=get_settings()['hr_draws_rule']
-        )
-        svg_str = fit_info.gen_bouts_graph_svg()
-        
-    except RuntimeError as e:
-        logMessage('horseRace_get_bouts_graph exception: %s' % str(e))
-    result = {'image': svg_str,
-              'announce_html': f"Bouts for {tourney.name}"
-              }
-
-    return result
-    
 
 def _tourney_json_rep(db, tourney):
     owner = db.query(User).filter_by(id=tourney.owner).one()
@@ -1017,7 +597,7 @@ def ajax_tourneys_settings(**kwargs):
                 db.add(tourney)
                 db.commit()
             else:
-                logMessage(f'The tournament {tourney.name} was not changed')
+                LOGGER.info(f'The tournament {tourney.name} was not changed')
             return {'status': 'success',
                     'value': json_rep
                     }
@@ -1054,7 +634,8 @@ def ajax_tourneys(**kwargs):
         elif request.method == 'PUT':
             assert 'action' in request.values, 'action is required for PUT requests'
             if request.values['action'] == 'delete':
-                assert 'tourney_id' in request.values, 'tourney_id is required for PUT "delete" requests'
+                assert 'tourney_id' in request.values, ('tourney_id is required for PUT'
+                                                        ' "delete" requests')
                 tourney_id = int(request.values['tourney_id'])
                 tourney = db.query(Tourney).filter_by(tourneyId=tourney_id).one()
                 check_can_delete(tourney)
@@ -1071,7 +652,8 @@ def ajax_tourneys(**kwargs):
                         'msg': f'{exc}'
                     }
             elif request.values['action'] == 'create':
-                assert 'tourney_id' not in request.values, 'tourney_id is forbidden for PUT "create" requests'
+                assert 'tourney_id' not in request.values, ('tourney_id is forbidden for'
+                                                            ' PUT "create" requests')
                 assert 'name' in request.values, 'name is required for PUT "create" requests'
                 name = request.values['name']
                 assert 'note' in request.values, 'note is required for PUT "create" requests'
@@ -1083,12 +665,13 @@ def ajax_tourneys(**kwargs):
                         'status': 'failure',
                         'msg': f"{exc}"
                         }
-                db.commit();
+                db.commit()
                 return {
                     'status': 'success',
                     'value': { 'tourney_id': tourney.tourneyId }
                 }
             else:
+                action = request.values['action']
                 return {
                     'status': 'failure',
                     'msg': f'unknown action "{action}" was requested'
@@ -1155,7 +738,7 @@ def ajax_entrants_settings(**kwargs):
                 db.add(player)
                 db.commit()
             else:
-                logMessage(f'The player {player.name} was not changed')
+                LOGGER.info(f'The player {player.name} was not changed')
             return {'status': 'success',
                     'value': json_rep
                     }
@@ -1173,19 +756,19 @@ def ajax_entrants_settings(**kwargs):
 def ajax_settings(**kwargs):
     if request.method == 'GET':
         return {'status': 'success',
-                'value': {k:v for k, v in get_settings().items()}
+                'value': dict(get_settings().items())
                 }
     elif request.method == 'PUT':
         name = request.values['name']
-        id = request.values['id']
+        setting_id = request.values['id']
         if name in get_settings():
             try:
-                set_settings(name, id);
+                set_settings(name, setting_id)
                 get_db().commit()
                 return {'status': 'success'}
-            except SettingsError as e:
+            except SettingsError:
                 return {"status": "error",
-                        "msg": f"invalid setting {name} = {id}"
+                        "msg": f"invalid setting {name} = {setting_id}"
                 }
         else:
             return {"status": "error",
@@ -1246,7 +829,8 @@ def ajax_entrants(**kwargs):
                     'value': {}
                 }
             elif request.values['action'] == 'delete':
-                assert 'player_id' in request.values, 'player_id is required for PUT "delete" requests'
+                assert 'player_id' in request.values, ('player_id is required for'
+                                                       ' PUT "delete" requests')
                 player_id = int(request.values['player_id'])
                 player = db.query(LogitPlayer).filter_by(id=player_id).one()
                 tourney.remove_player(db, player)
@@ -1256,7 +840,8 @@ def ajax_entrants(**kwargs):
                     'value': {}
                 }
             elif request.values['action'] == 'create':
-                assert 'player_id' not in request.values, 'player_id is forbidden for PUT "create" requests'
+                assert 'player_id' not in request.values, ('player_id is forbidden for'
+                                                           ' PUT "create" requests')
                 assert 'name' in request.values, 'name is required for PUT "create" requests'
                 name = request.values['name']
                 assert 'note' in request.values, 'note is required for PUT "create" requests'
@@ -1268,14 +853,15 @@ def ajax_entrants(**kwargs):
                         'status': 'failure',
                         'msg': f"{exc}"
                         }
-                db.commit();
-                tourney.add_player(db, player);
-                db.commit();
+                db.commit()
+                tourney.add_player(db, player)
+                db.commit()
                 return {
                     'status': 'success',
                     'value': { 'player_id': player.id }
                 }
             else:
+                action = request.values['action']
                 return {
                     'status': 'failure',
                     'msg': f'unknown action "{action}" was requested'
@@ -1320,7 +906,6 @@ def ajax_bouts_settings(**kwargs):
             check_can_read(lplayer)
             check_can_read(rplayer)
             json_rep = bout.as_dict()
-            changed = 0
             change_dct = {}
             for key in json_rep:
                 # The request comes back with player ids in the name slots;
@@ -1338,7 +923,7 @@ def ajax_bouts_settings(**kwargs):
                 db.add(bout)
                 db.commit()
             else:
-                logMessage(f'Bout {bout.boutId} was not changed')
+                LOGGER.info(f'Bout {bout.boutId} was not changed')
             json_rep['lplayer'] = bout.lName
             json_rep['rplayer'] = bout.rName
             return {'status': 'success',
@@ -1387,8 +972,8 @@ def ajax_bouts(**kwargs):
             if request.values['action'] == 'add':
                 for key in ['tourney_id', 'lplayer', 'rplayer']:
                     assert key in request.values, f'{key} is required for PUT "add" requests'
-                if not any([key in request.values for key in ['lwins', 'rwins', 'draws',
-                                                              'note']]):
+                if not any(key in request.values for key in ['lwins', 'rwins', 'draws',
+                                                             'note']):
                     raise AssertionError('At least one of lwins, rwins, draws, or note'
                                          'is required for PUT "add" requests')
                 tourney_id = int(request.values['tourney_id'])
@@ -1418,7 +1003,7 @@ def ajax_bouts(**kwargs):
                 }
             elif request.values['action'] == 'delete':
                 assert 'bout_id' in request.values, 'bout_id is required for PUT "delete" requests'
-                bout_id = int(request.values['bout_id']);
+                bout_id = int(request.values['bout_id'])
                 bout = db.query(Bout).filter_by(boutId=bout_id).one()
                 tourney = db.query(Tourney).filter_by(tourneyId=bout.tourneyId).one()
                 check_can_write(tourney)
@@ -1434,6 +1019,7 @@ def ajax_bouts(**kwargs):
         return {'status': 'failure',
                 'msg': f"{excinfo}"
                 }
+    raise RuntimeError("logic error - this should not be reached")
 
 
 @bp.route('/ajax/bearpit', methods=["GET", "PUT"])
@@ -1511,8 +1097,8 @@ def ajax_wfw(**kwargs):
                     draws_rule=get_settings()['hr_draws_rule']
                 )
                 svg_str = fit_info.gen_bouts_graph_svg()
-            except RuntimeError as e:
-                logMessage('ajax_wfw exception: %s' % str(e))
+            except RuntimeError as excinfo:
+                LOGGER.info(f'ajax_wfw exception: {excinfo}')
                 raise
             response_data = {
                 'image': svg_str,
@@ -1569,8 +1155,8 @@ def ajax_whoiswinning(**kwargs):
                 svg_str = fit_info.gen_horserace_graph_svg(graph_yscale, graph_type)
                 announce_html = fit_info.estimate_win_probabilities().as_html()
 
-            except RuntimeError as e:
-                logMessage('ajax_whoiswinning exception: %s' % str(e))
+            except RuntimeError as excinfo:
+                LOGGER.info(f'ajax_whoiswinning exception: {excinfo}')
                 raise
             response_data = {
                 'image': svg_str,
@@ -1609,7 +1195,7 @@ def _set_hr_include(tourney_id, player_id, flag: bool):
     tourney_dct[str(player_id)] = flag
     dct[str(tourney_id)] = tourney_dct
     session['horserace_includes'] = dct
-    
+
 
 @bp.route('/ajax/horserace/checkbox', methods=["GET", "PUT"])
 @debug_wrapper
@@ -1639,150 +1225,3 @@ def handle_horserace_checkbox(**kwargs):
                 }
     else:
         return f"unsupported method {request.method}", 405
-
-
-@bp.route('/json/<path>')
-@debug_wrapper
-def handleJSON(path, **kwargs):
-    """
-    Specialized endpoint called by jqgrid when populating tables
-    """
-    db = get_db()
-    engine = db.get_bind()
-    if path=='tourneys':
-        tourneyList = get_readable_tourneys(db)
-        result = {
-                  "records":len(tourneyList),  # total records
-                  "rows": [ {"id":t.tourneyId,
-                             "cell":[t.tourneyId,
-                                     t.name,
-                                     t.ownerName or '-nobody-',
-                                     t.groupName or '-no group-',
-                                     t.note,
-                                     t.tourneyId]}
-                           for t in tourneyList ]
-                  }
-    elif path=='entrants':
-        tourneyId = int(request.values.get('tourneyId', -1))
-        if tourneyId != -1:
-            # we are dealing with a specific tourney
-            tourney = (db.query(Tourney)
-                       .filter_by(tourneyId=tourneyId)
-                       .one()
-                       )
-            check_can_read(tourney)
-        session['sel_tourney_id'] = tourneyId
-        with engine.connect() as conn:
-            rs1 = conn.execute(  # This gets the players with bouts
-                """
-                select players.name as name, players.id as id,
-                  count(distinct bouts.tourneyId) as num_tournies,
-                  count(boutId) as num_bouts, players.note as note
-                from players, bouts
-                where 
-                  players.id = bouts.leftPlayerId
-                  or players.id = bouts.rightPlayerId
-                group by players.id
-                """
-            )
-            if tourneyId >= 0:
-                stxt = sql_text(
-                    """
-                    select players.id as id
-                    from players, bouts
-                    where
-                      bouts.tourneyId = :t_id
-                      and (players.id = bouts.leftPlayerId
-                           or players.id = bouts.rightPlayerId)
-                    """
-                )
-                player_rs = conn.execute(stxt, t_id=tourneyId)
-                player_id_set = set([val.id for val in player_rs])
-                playerList = [val for val in rs1 if val.id in player_id_set]
-            else:
-                rs2 = conn.execute(  # This gets the players without bouts
-                    """
-                    select players.name as name, players.id as id,
-                      0 as num_tournies, 0 as num_bouts, players.note as note
-                    from players, bouts
-                    where 
-                      players.id not in (select leftPlayerId from bouts)
-                      and
-                      players.id not in (select rightPlayerId from bouts)
-                    group by players.id
-                    """
-                )
-                playerList = ([val for val in rs1] + [val for val in rs2])
-            result = {
-                      "records":len(playerList),  # total records
-                      "rows": [ {"id":p.id,
-                                 "cell":[p.id, p.name,
-                                         p.num_bouts, p.num_tournies,
-                                         p.note]} for p in playerList ]
-                      }
-    elif path =='bouts':
-        tourneyId = int(request.values.get('tourneyId', -1))
-        session['sel_tourney_id'] = tourneyId
-        if tourneyId >= 0:
-            tourney = (db.query(Tourney)
-                       .filter_by(tourneyId=tourneyId)
-                       .one()
-                       )
-            check_can_read(tourney)
-            bout_list = [b for b in (db.query(Bout)
-                                     .filter_by(tourneyId=tourneyId)
-                                     .all())]
-        else:
-            tourney_list = get_readable_tourneys(db)
-            tourney_id_list = [tourney.tourneyId
-                               for tourney in tourney_list]
-            bout_list = (db.query(Bout)
-                         .filter(Bout.tourneyId.in_(tourney_id_list))
-                         .all()
-                         )
-            bout_id_list = [b.boutId for b in list(bout_list)]
-        result = {
-                  "records":len(bout_list),  # total records
-                  "rows": [ {"id":p.boutId, "cell":[p.tourneyName, 
-                                                    p.leftWins, p.lName, p.draws,
-                                                    p.rName, p.rightWins, p.note] } 
-                            for p in bout_list]
-                  }
-    elif path == 'horserace':
-        paramList = ['%s:%s'%(str(k),str(v)) for k,v in list(request.values.items())]
-        tourneyId = int(request.values['tourney'])
-        session['sel_tourney_id'] = tourneyId
-        if tourneyId >= 0:
-            tourney = (db.query(Tourney)
-                       .filter_by(tourneyId=tourneyId)
-                       .one()
-                       )
-            check_can_read(tourney)
-            bouts = tourney.get_bouts(db)
-        else:
-            tourney_list = get_readable_tourneys(db)
-            tourney_id_list = [tourney.tourneyId
-                               for tourney in tourney_list]
-            bouts = (db.query(Bout)
-                     .filter(Bout.tourneyId.in_(tourney_id_list))
-                     .all()
-                     )
-
-        workDF = _get_bearpit_dataframe(db, bouts)
-
-        print(workDF)
-            
-        result = {
-                  "records":len(workDF),  # total records
-                  "rows": [{"id":row['id'],
-                            "cell":[row['id'], row['playerName'],
-                                    row['wins'], row['losses'], row['draws'],
-                                    row['bearpit'],
-                                    str(row['id']) + ('+' if _get_hr_include(tourneyId,
-                                                                             row['id'])
-                                                      else '-')]} 
-                            for _,row in workDF.iterrows() ]
-                  }
-    else:
-        raise RuntimeError("Request for unknown AJAX element %s"%path)
-    return result
