+++ src/bayestourney/views/base.html
@@ -72,10 +72,10 @@
           </button>
           <div class="collapse navbar-collapse justify-content-end" id="navbarSupportedContent">
             <ul class="nav navbar-nav me-auto mb-2 mb-lg-0">
-              {{ nav_link('tourneys', 'Tournaments') }}
-              {{ nav_link('entrants', 'Entrants') }}
-              {{ nav_link('bouts', 'Bouts') }}
-              {{ nav_link('horserace', 'Horserace') }}
+              {{ nav_link('tourneys_fun', 'Tournaments') }}
+              {{ nav_link('entrants_fun', 'Entrants') }}
+              {{ nav_link('bouts_fun', 'Bouts') }}
+              {{ nav_link('horserace_fun', 'Horserace') }}
             </ul>
             {#
             <form class="d-flex">
@@ -88,7 +88,7 @@
               {{ nav_link('admin_page', 'Admin') }}
               {% endif %}
               {{ nav_link('settings', 'Settings') }}
-              {{ nav_link('help', 'Help') }}
+              {{ nav_link('help_fun', 'Help') }}
               <li class="nav-item dropdown">
                 <a href="#" class="nav-link dropdown-toggle" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                   Misc <b class="caret"></b></a>
