+++ src/bayestourney/views/notimpl.html
@@ -1,0 +1,9 @@
+{% extends 'base.html' %}
+
+{% block pagescripts %}
+{% endblock %}
+
+{% block content %}
+<h1>Not Implemented</h1>
+You have requested content which has not yet been implemented.
+{% endblock %}
